moy_grisha/requirements.txt
amqp==2.6.0
asgiref==3.2.10
billiard==3.6.3.0
boto3==1.14.12
botocore==1.17.12
celery==4.4.6
certifi==2020.6.20
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.7
django-countries==6.1.2
django-crispy-forms==1.9.1
django-embed-video==1.3.3
django-heroku==0.3.1
django-ordered-model==3.4.1
django-paypal==1.0.0
django-plans==0.8.13
django-plans-paypal==0.0.2
django-storages==1.9.1
docutils==0.15.2
future==0.18.2
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.7.0
jmespath==0.10.0
kombu==4.6.11
paypal==1.2.5
Pillow==7.1.2
psycopg2==2.8.5
python-dateutil==2.8.1
python-git==2018.2.1
python-stdnum==1.13
pytz==2020.1
requests==2.24.0
s3transfer==0.3.3
Send2Trash==1.5.0
six==1.15.0
sqlparse==0.3.1
suds-jurko==0.6
urllib3==1.25.9
vatnumber==1.2
vine==1.3.0
whitenoise==5.1.0
zipp==3.1.0

